backend/requirements.txt
# FastAPI
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# DB
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9
# 설정
pydantic-settings==2.2.1
python-dotenv==1.0.1
# Celery
celery[redis]==5.3.6
redis==5.0.4
# 수집
httpx==0.27.0
feedparser==6.0.11
aiohttp==3.9.5
Telethon==1.37.0
# NLP (Week 2)
langdetect==1.0.9
deep-translator>=1.11.4
# 보안
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9
sentry-sdk[fastapi]==2.14.0
firebase-admin>=6.3.0
# Google Play Billing
google-auth>=2.29.0
google-api-python-client>=2.127.0
# Apple StoreKit (JWT 생성)
PyJWT[crypto]>=2.8.0
cryptography>=42.0.0
# 유틸
geohash2==1.1
tenacity==8.3.0
structlog==24.1.0
# 테스트
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==5.0.0
fakeredis==2.23.1
factory-boy==3.3.0
